chore(wallet): remove deprecated components
