+++ app/src/main/kotlin/eu/peernetwork/app/ui/feed/FeedModule.kt
@@ -24,7 +24,7 @@ import eu.peernetwork.core.ui.factory.UiViewModelFactory
 import eu.peernetwork.social.ui.connection.Connection
 import eu.peernetwork.wallet.ui.balance.Balance
 import eu.peernetwork.wallet.ui.confirmation.Confirmation
-import eu.peernetwork.wallet.ui.dashboard.Dashboard
+import eu.peernetwork.wallet.ui.rate.Rate
 import javax.inject.Provider
 
 @Module
@@ -129,9 +129,9 @@ object FeedModule {
     @Feed.Scope
     @Provides
     @IntoMap
-    @UiBuilder(Dashboard.Builder::class)
+    @UiBuilder(Rate.Builder::class)
     fun provideServiceBuilder(component: Feed.Component): UiComponent.Builder {
-        return Dashboard.Builder(component)
+        return Rate.Builder(component)
     }
 
     @Provides
