+++ app/src/main/kotlin/eu/peernetwork/app/ui/profile/Profile.kt
@@ -21,7 +21,7 @@ import eu.peernetwork.social.ui.report.Report
 import eu.peernetwork.user.ui.user.User
 import eu.peernetwork.wallet.ui.balance.Balance
 import eu.peernetwork.wallet.ui.confirmation.Confirmation
-import eu.peernetwork.wallet.ui.dashboard.Dashboard as WalletDashboard
+import eu.peernetwork.wallet.ui.rate.Rate as WalletDashboard
 
 interface Profile : ApplicationProvider {
     fun settingsEvent(): SettingsEvent
