+++ app/src/main/kotlin/eu/peernetwork/app/ui/search/Search.kt
@@ -17,7 +17,7 @@ import eu.peernetwork.social.ui.search.tag.Tag
 import eu.peernetwork.social.ui.search.title.Title
 import eu.peernetwork.wallet.ui.balance.Balance
 import eu.peernetwork.wallet.ui.confirmation.Confirmation
-import eu.peernetwork.wallet.ui.dashboard.Dashboard
+import eu.peernetwork.wallet.ui.rate.Rate
 
 interface Search : ApplicationProvider {
     fun settingsEvent(): SettingsEvent
@@ -42,7 +42,7 @@ interface Search : ApplicationProvider {
         Connection,
         Confirmation,
         Boost,
-        Dashboard,
+        Rate,
         Balance,
         Checkout,
         Screen
