+++ app/src/main/kotlin/eu/peernetwork/app/ui/wallet/WalletModule.kt
@@ -11,7 +11,7 @@ import eu.peernetwork.core.ui.component.UiComponentProvider
 import eu.peernetwork.core.ui.factory.UiBuilderFactory
 import eu.peernetwork.social.ui.search.member.Member
 import eu.peernetwork.wallet.ui.balance.Balance
-import eu.peernetwork.wallet.ui.dashboard.Dashboard
+import eu.peernetwork.wallet.ui.rate.Rate
 import eu.peernetwork.wallet.ui.transactions.Transactions
 import eu.peernetwork.wallet.ui.transfer.Transfer
 import javax.inject.Provider
@@ -44,9 +44,9 @@ object WalletModule {
     @Provides
     @Wallet.Scope
     @IntoMap
-    @UiBuilder(Dashboard.Builder::class)
+    @UiBuilder(Rate.Builder::class)
     fun provideServiceBuilder(component: Wallet.Component): UiComponent.Builder {
-        return Dashboard.Builder(component)
+        return Rate.Builder(component)
     }
 
     @Wallet.Scope
