--- wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/dashboard/DashboardModule.kt
@@ -1,47 +1,0 @@
-package eu.peernetwork.wallet.ui.dashboard
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import dagger.Module
-import dagger.Provides
-import dagger.multibindings.IntoMap
-import eu.peernetwork.core.ui.annotation.UiBuilder
-import eu.peernetwork.core.ui.annotation.UiViewModel
-import eu.peernetwork.core.ui.component.UiComponent
-import eu.peernetwork.core.ui.component.UiComponentProvider
-import eu.peernetwork.core.ui.factory.UiBuilderFactory
-import eu.peernetwork.core.ui.factory.UiViewModelFactory
-import eu.peernetwork.wallet.ui.transfer.Transfer
-import javax.inject.Provider
-
-@Module
-object DashboardModule {
-    @Provides
-    @Dashboard.Scope
-    fun provideBuilderFactory(factory: Map<Class<out UiComponent.Builder>,
-            @JvmSuppressWildcards Provider<UiComponent.Builder>>): UiComponentProvider.Factory {
-        return UiBuilderFactory(factory)
-    }
-
-    @Provides
-    @Dashboard.Scope
-    fun provideViewModelFactory(
-        classToViewModel: @JvmSuppressWildcards Map<Class<out ViewModel>, Provider<ViewModel>>
-    ): ViewModelProvider.Factory {
-        return UiViewModelFactory(classToViewModel)
-    }
-
-    @Provides
-    @IntoMap
-    @Dashboard.Scope
-    @UiViewModel(DashboardViewModel::class)
-    fun viewModel(viewModel: DashboardViewModel): ViewModel = viewModel
-
-    @Provides
-    @Dashboard.Scope
-    @IntoMap
-    @UiBuilder(Transfer.Builder::class)
-    fun provideTransferBuilder(component: Dashboard.Component): UiComponent.Builder {
-        return Transfer.Builder(component)
-    }
-}
