--- wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/dashboard/DashboardTransfer.kt
@@ -1,74 +1,0 @@
-package eu.peernetwork.wallet.ui.dashboard
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.res.stringResource
-import androidx.lifecycle.ViewModelStoreOwner
-import eu.peernetwork.wallet.ui.R
-import eu.peernetwork.wallet.ui.compose.ExpandableLabel
-import eu.peernetwork.wallet.ui.model.UiTransfer
-import eu.peernetwork.wallet.ui.transfer.TransferScreen
-import eu.peernetwork.wallet.ui.transfer.TransferSheet
-
-@Composable
-fun DashboardTransfer(
-    tax: Int,
-    state: MutableState<DashboardState>,
-    component: Dashboard.Component,
-    viewModelStoreOwner: ViewModelStoreOwner,
-    onAccountClicked: (String) -> Unit,
-    onClear: () -> Unit,
-    onClick: () -> Unit,
-) {
-    val lastState = remember(state.value) { mutableStateOf(state.value) }
-    val showTransfer = rememberSaveable { mutableStateOf(false) }
-    val transaction = remember { mutableStateOf<UiTransfer?>(null) }
-    val handleOnClear by rememberUpdatedState(onClear)
-    val handleOnAccountClicked by rememberUpdatedState(onAccountClicked)
-    ExpandableLabel(
-        stringResource(R.string.transfer_label),
-        showTransfer,
-        {
-            if (lastState.value is DashboardState.Default && !it) {
-                handleOnClear()
-            }
-            state.value = lastState.value
-        }
-    ) {
-        DashboardPage(
-            stringResource(R.string.recipient_selection_label),
-            state,
-            onClick,
-        ) {
-            TransferScreen(
-                tax = tax,
-                recipient = it.recipient,
-                provider = component,
-                viewModelStoreOwner = viewModelStoreOwner,
-                onClear = {
-                    lastState.value = DashboardState.Default
-                    showTransfer.value = false },
-                onTransfer = { transaction.value = it }
-            ) { handleOnAccountClicked(it.id) }
-        }
-    }
-    (state.value as? DashboardState.Transfer?)?.let {
-        TransferSheet(
-            transaction,
-            component,
-            viewModelStoreOwner,
-            it.recipient,
-            {
-                transaction.value = null
-                lastState.value = DashboardState.Default
-                showTransfer.value = false
-            }
-        ) { handleOnAccountClicked(it.id) }
-    }
-}
-
