+++ wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/rate/Rate.kt
@@ -1,30 +1,29 @@
-package eu.peernetwork.wallet.ui.dashboard
+package eu.peernetwork.wallet.ui.rate
 
 import android.content.Context
 import androidx.lifecycle.ViewModelProvider
 import eu.peernetwork.core.common.provider.CoreProvider
 import eu.peernetwork.core.ui.component.UiComponent
-import eu.peernetwork.core.ui.component.UiComponentProvider
 import eu.peernetwork.wallet.ui.provider.WalletProvider
 import eu.peernetwork.wallet.ui.transfer.Transfer
 
-interface Dashboard : WalletProvider, CoreProvider {
+interface Rate : WalletProvider, CoreProvider {
     @javax.inject.Scope
     @Retention(AnnotationRetention.RUNTIME)
     annotation class Scope
 
     @Scope
     @dagger.Component(
-        dependencies = [ Dashboard::class ],
-        modules = [ DashboardModule::class ]
+        dependencies = [ Rate::class ],
+        modules = [ RateModule::class ]
     )
-    interface Component : Dashboard, Transfer, UiComponentProvider {
+    interface Component : Rate, Transfer {
         fun viewModelFactory(): ViewModelProvider.Factory
     }
 
-    class Builder(private val dependency: Dashboard): UiComponent.DefaultBuilder<Dashboard, Component>() {
+    class Builder(private val dependency: Rate): UiComponent.DefaultBuilder<Rate, Component>() {
         override fun build(context: Context): Component {
-            return DaggerDashboard_Component.builder().dashboard(dependency).build()
+            return DaggerRate_Component.builder().rate(dependency).build()
         }
     }
 }
