+++ wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/transfer/TransferMessage.kt
@@ -1,4 +1,4 @@
-package eu.peernetwork.wallet.ui.transfer.v2
+package eu.peernetwork.wallet.ui.transfer
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
