+++ wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/transfer/TransferPreview.kt
@@ -1,4 +1,4 @@
-package eu.peernetwork.wallet.ui.transfer.v2
+package eu.peernetwork.wallet.ui.transfer
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
