+++ wallet/ui/src/main/kotlin/eu/peernetwork/wallet/ui/transfer/TransferSuccess.kt
@@ -1,6 +1,5 @@
-package eu.peernetwork.wallet.ui.transfer.v2
+package eu.peernetwork.wallet.ui.transfer
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -13,20 +12,25 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.animateLottieCompositionAsState
+import com.airbnb.lottie.compose.rememberLottieComposition
 import eu.peernetwork.core.ui.design.luna.DesignButton
 import eu.peernetwork.core.ui.design.material.DesignBottomSheetScaffold
 import eu.peernetwork.core.ui.theme.DesignTheme
@@ -72,10 +76,8 @@ fun TransferSuccess(
                 .clip(CircleShape)
                 .background(PeerAppLightGreen.copy(alpha = .1f))
         ) {
-            Image(
-                painter = painterResource(R.drawable.ic_successful),
-                contentDescription = stringResource(R.string.transaction_status),
-                modifier = Modifier
+            TransferSuccessIcon(
+                Modifier
                     .padding(8.dp)
                     .fillMaxSize()
             )
@@ -100,6 +102,28 @@ fun TransferSuccess(
                 .padding(bottom = 6.dp)
         ) { Text(stringResource(R.string.got_it_label)) }
     }
+}
+
+@Composable
+fun TransferSuccessIcon(
+    modifier: Modifier = Modifier,
+    rawRes: Int = R.raw.tick
+) {
+    var playing by remember { mutableStateOf(false) }
+    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(rawRes))
+    val progress by animateLottieCompositionAsState(
+        composition = composition,
+        isPlaying = playing,
+        reverseOnRepeat = true,
+        iterations = 1,
+        speed = 1f
+    )
+    LaunchedEffect(Unit) { playing = true }
+    LottieAnimation(
+        composition = composition,
+        progress = { progress },
+        modifier = modifier
+    )
 }
 
 @Preview
