+++ wallet/ui/src/test/kotlin/eu/peernetwork/wallet/ui/service/ServiceViewModelTest.kt
@@ -4,7 +4,7 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import app.cash.turbine.test
 import eu.peernetwork.wallet.domain.model.Tax
 import eu.peernetwork.wallet.domain.usecase.TaxUsecase
-import eu.peernetwork.wallet.ui.dashboard.DashboardViewModel
+import eu.peernetwork.wallet.ui.rate.RateViewModel
 import eu.peernetwork.wallet.ui.mapper.mapFromDomain
 import io.mockk.coEvery
 import io.mockk.mockk
@@ -30,12 +30,12 @@ internal class ServiceViewModelTest {
 
     private val usecase = mockk<TaxUsecase>()
 
-    private lateinit var viewModel: DashboardViewModel
+    private lateinit var viewModel: RateViewModel
 
     @Before
     fun setup() {
         Dispatchers.setMain(dispatcher)
-        viewModel = DashboardViewModel(usecase)
+        viewModel = RateViewModel(usecase)
     }
 
     @After
@@ -57,8 +57,8 @@ internal class ServiceViewModelTest {
         }
         viewModel.initialize()
         viewModel.state.test {
-            assertEquals(DashboardViewModel.State.Loading, awaitItem())
-            assertEquals(DashboardViewModel.State.Success(tax.mapFromDomain()), awaitItem())
+            assertEquals(RateViewModel.State.Loading, awaitItem())
+            assertEquals(RateViewModel.State.Success(tax.mapFromDomain()), awaitItem())
         }
     }
 
@@ -68,7 +68,7 @@ internal class ServiceViewModelTest {
         coEvery { usecase() } throws exception
         viewModel.initialize()
         viewModel.state.test {
-            assertEquals(DashboardViewModel.State.Error(exception), awaitItem())
+            assertEquals(RateViewModel.State.Error(exception), awaitItem())
         }
     }
 }
